requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
antiorm==1.2.1
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.6
db==0.1.1
db-sqlite3==0.0.1
Django==4.2.29
django-allauth==65.15.0
djangorestframework==3.17.1
frozenlist==1.8.0
gunicorn==25.3.0
idna==3.11
jwt==1.4.0
multidict==6.7.1
nodejs-wheel-binaries==24.14.0
packaging==26.0
pillow==12.1.1
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==3.0
python-decouple==3.8
pytube==15.0.0
pytubefix==10.3.8
pytz==2026.1.post1
requests==2.33.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
yarl==1.23.0
